+++ src/backend/Application.API/SaveFiles/List/ListSaveFilesQuery.cs
@@ -2,7 +2,7 @@
 
 namespace Application.SaveFiles.List;
 
-public sealed record ListSaveFilesQuery(Guid UserId, uint Offset, ushort Limit) : IQuery<ListSaveFilesResult>
+public sealed record ListSaveFilesQuery(Guid UserId, ushort Offset, ushort Limit) : IQuery<ListSaveFilesResult>
 {
     public const string Path = "/savefiles";
 }
